requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
async-timeout==5.0.1
beautifulsoup4==4.14.2
certifi==2025.10.5
charset-normalizer==3.4.4
colorama==0.4.6
distro==1.9.0
Django==5.2.7
django-autoslug==1.9.9
django-ckeditor==6.7.3
django-classy-tags==4.1.0
django-easy-seo==0.4.8
django-grappelli==4.0.2
django-jazzmin==3.0.1
django-js-asset==3.1.2
django-meta==2.5.0
django-redis==6.0.0
django-robots==6.1
django-semantic-admin==0.6.6
django-semantic-forms==0.1.8
django-seo2==1.0.1
exceptiongroup==1.3.0
filelock==3.20.0
fsspec==2025.9.0
gradio_client==1.13.3
gunicorn==23.0.0
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
idna==3.11
jiter==0.11.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.1.0
langchain-core==1.1.0
langchain-openai==1.1.0
langgraph==1.0.3
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.2.10
langsmith==0.4.47
lxml==6.0.2
openai==2.8.1
orjson==3.11.4
ormsgpack==1.12.0
packaging==25.0
pillow==12.0.0
psycopg==3.2.10
psycopg-binary==3.2.10
pydantic==2.12.3
pydantic_core==2.41.4
PyYAML==6.0.3
redis==7.0.1
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
sniffio==1.3.1
soupsieve==2.8
sqlparse==0.5.3
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
whitenoise==6.11.0
xxhash==3.6.0
zstandard==0.25.0
dj-database-url